init: move all init related stuff from cpu.c to init.c
