+++ hypervisor/arch/x86/cpu.c
@@ -5,33 +5,21 @@
  */
 
 #include <types.h>
+#include <logmsg.h>
+#include <udelay.h>
+#include <board_info.h>
 #include <x86/lib/bits.h>
-#include <x86/page.h>
-#include <x86/e820.h>
-#include <x86/mmu.h>
-#include <x86/vtd.h>
-#include <x86/lapic.h>
 #include <x86/per_cpu.h>
+#include <x86/cpuid.h>
+#include <x86/lapic.h>
 #include <x86/cpufeatures.h>
 #include <x86/cpu_caps.h>
-#include <acpi.h>
-#include <x86/ioapic.h>
-#include <x86/trampoline.h>
-#include <x86/cpuid.h>
-#include <version.h>
+#include <x86/irq.h>
+#include <x86/mmu.h>
 #include <x86/vmx.h>
 #include <x86/msr.h>
-#include <ptdev.h>
-#include <logmsg.h>
-#include <x86/rdt.h>
-#include <x86/sgx.h>
-#include <uart16550.h>
-#include <vpci.h>
-#include <ivshmem.h>
-#include <x86/rtcm.h>
-#include <x86/tsc.h>
-#include <udelay.h>
-#include <cycles.h>
+#include <x86/page.h>
+#include <x86/trampoline.h>
 
 #define CPU_UP_TIMEOUT		100U /* millisecond */
 #define CPU_DOWN_TIMEOUT	100U /* millisecond */
@@ -44,33 +32,7 @@ static uint64_t startup_paddr = 0UL;
 /* physical cpu active bitmap, support up to 64 cpus */
 static volatile uint64_t pcpu_active_bitmap = 0UL;
 
-static void init_pcpu_xsave(void);
 static void set_current_pcpu_id(uint16_t pcpu_id);
-static void print_hv_banner(void);
-static uint16_t get_pcpu_id_from_lapic_id(uint32_t lapic_id);
-static uint64_t start_cycle __attribute__((__section__(".bss_noinit")));
-
-/**
- * @pre phys_cpu_num <= MAX_PCPU_NUM
- */
-static bool init_percpu_lapic_id(void)
-{
-	uint16_t i;
-	uint32_t lapic_id_array[MAX_PCPU_NUM];
-	bool success = false;
-
-	/* Save all lapic_id detected via parse_mdt in lapic_id_array */
-	phys_cpu_num = parse_madt(lapic_id_array);
-
-	if ((phys_cpu_num != 0U) && (phys_cpu_num <= MAX_PCPU_NUM)) {
-		for (i = 0U; i < phys_cpu_num; i++) {
-			per_cpu(lapic_id, i) = lapic_id_array[i];
-		}
-		success = true;
-	}
-
-	return success;
-}
 
 static void pcpu_set_current_state(uint16_t pcpu_id, enum pcpu_boot_state state)
 {
@@ -85,12 +47,26 @@ static void pcpu_set_current_state(uint16_t pcpu_id, enum pcpu_boot_state state)
 	per_cpu(boot_state, pcpu_id) = state;
 }
 
+/**
+ * @pre num <= MAX_PCPU_NUM
+ */
+void set_pcpu_nums(uint16_t num)
+{
+	phys_cpu_num = num;
+}
+
 /*
  * @post return <= MAX_PCPU_NUM
  */
 uint16_t get_pcpu_nums(void)
 {
 	return phys_cpu_num;
+}
+
+static void set_active_pcpu_bitmap(uint16_t pcpu_id)
+{
+	bitmap_set_lock(pcpu_id, &pcpu_active_bitmap);
+
 }
 
 bool is_pcpu_active(uint16_t pcpu_id)
@@ -103,220 +79,21 @@ uint64_t get_active_pcpu_bitmap(void)
 	return pcpu_active_bitmap;
 }
 
-static void enable_ac_for_splitlock(void)
+void init_pcpu_state(uint16_t pcpu_id)
 {
-#ifndef CONFIG_ENFORCE_TURNOFF_AC
-	uint64_t test_ctl;
-
-	if (has_core_cap(1U << 5U)) {
-		test_ctl = msr_read(MSR_TEST_CTL);
-		test_ctl |= (1U << 29U);
-		msr_write(MSR_TEST_CTL, test_ctl);
-	}
-#endif /*CONFIG_ENFORCE_TURNOFF_AC*/
-}
-
-void init_pcpu_pre(bool is_bsp)
-{
-	uint16_t pcpu_id;
-	int32_t ret;
-
-	if (is_bsp) {
-		pcpu_id = BSP_CPU_ID;
-		start_cycle = get_cpu_cycles();
-
-		/* Get CPU capabilities thru CPUID, including the physical address bit
-		 * limit which is required for initializing paging.
-		 */
-		init_pcpu_capabilities();
-
-		if (detect_hardware_support() != 0) {
-			panic("hardware not support!");
-		}
-
-		init_pcpu_model_name();
-
-		load_pcpu_state_data();
-
-		/* Initialize the hypervisor paging */
-		init_e820();
-		init_paging();
-
-		/*
-		 * Need update uart_base_address here for vaddr2paddr mapping may changed
-		 * WARNNING: DO NOT CALL PRINTF BETWEEN ENABLE PAGING IN init_paging AND HERE!
-		 */
-		uart16550_init(false);
-
-		early_init_lapic();
-
-#ifdef CONFIG_ACPI_PARSE_ENABLED
-		ret = acpi_fixup();
-		if (ret != 0) {
-			panic("failed to parse/fix up ACPI table!");
-		}
-#endif
-
-		if (!init_percpu_lapic_id()) {
-			panic("failed to init_percpu_lapic_id!");
-		}
-
-		ret = init_ioapic_id_info();
-		if (ret != 0) {
-			panic("System IOAPIC info is incorrect!");
-		}
-
-#ifdef CONFIG_RDT_ENABLED
-		init_rdt_info();
-#endif
-
-		/* NOTE: this must call after MMCONFIG is parsed in acpi_fixup() and before APs are INIT.
-		 * We only support platform with MMIO based CFG space access.
-		 * IO port access only support in debug version.
-		 */
-		pci_switch_to_mmio_cfg_ops();
-	} else {
-
-		/* Switch this CPU to use the same page tables set-up by the
-		 * primary/boot CPU
-		 */
-		enable_paging();
-
-		early_init_lapic();
-
-		pcpu_id = get_pcpu_id_from_lapic_id(get_cur_lapic_id());
-		if (pcpu_id >= MAX_PCPU_NUM) {
-			panic("Invalid pCPU ID!");
-		}
-	}
-
-	bitmap_set_lock(pcpu_id, &pcpu_active_bitmap);
+	set_active_pcpu_bitmap(pcpu_id);
 
 	/* Set state for this CPU to initializing */
 	pcpu_set_current_state(pcpu_id, PCPU_STATE_INITIALIZING);
 }
 
-void init_pcpu_post(uint16_t pcpu_id)
-{
-#ifdef STACK_PROTECTOR
-	set_fs_base();
-#endif
-	load_gdtr_and_tr();
-
-	enable_ac_for_splitlock();
-
-	init_pcpu_xsave();
-
-	if (pcpu_id == BSP_CPU_ID) {
-		/* Print Hypervisor Banner */
-		print_hv_banner();
-
-		/* Calibrate TSC Frequency */
-		calibrate_tsc();
-
-		pr_acrnlog("HV version %s-%s-%s %s (daily tag:%s) %s@%s build by %s%s, start time %luus",
-				HV_FULL_VERSION,
-				HV_BUILD_TIME, HV_BUILD_VERSION, HV_BUILD_TYPE,
-				HV_DAILY_TAG, HV_BUILD_SCENARIO, HV_BUILD_BOARD,
-				HV_BUILD_USER, HV_CONFIG_TOOL, cycles_to_us(start_cycle));
-
-		pr_acrnlog("API version %u.%u",	HV_API_MAJOR_VERSION, HV_API_MINOR_VERSION);
-
-		pr_acrnlog("Detect processor: %s", (get_pcpu_info())->model_name);
-
-		pr_dbg("Core %hu is up", BSP_CPU_ID);
-
-		/* Warn for security feature not ready */
-		if (!check_cpu_security_cap()) {
-			pr_fatal("SECURITY WARNING!!!!!!");
-			pr_fatal("Please apply the latest CPU uCode patch!");
-		}
-
-		/* Initialize interrupts */
-		init_interrupt(BSP_CPU_ID);
-
-		/* Setup ioapic irqs */
-		ioapic_setup_irqs();
-
-		timer_init();
-		setup_notification();
-		setup_pi_notification();
-
-		if (init_iommu() != 0) {
-			panic("failed to initialize iommu!");
-		}
-
-#ifdef CONFIG_IVSHMEM_ENABLED
-		init_ivshmem_shared_memory();
-#endif
-		init_pci_pdev_list(); /* init_iommu must come before this */
-		ptdev_init();
-
-		if (init_sgx() != 0) {
-			panic("failed to initialize sgx!");
-		}
-
-		/*
-		 * Reserve memory from platform E820 for EPT 4K pages for all VMs
-		 */
-#ifdef CONFIG_LAST_LEVEL_EPT_AT_BOOT
-		reserve_buffer_for_ept_pages();
-#endif
-		/* Start all secondary cores */
-		startup_paddr = prepare_trampoline();
-		if (!start_pcpus(AP_MASK)) {
-			panic("Failed to start all secondary cores!");
-		}
-
-		ASSERT(get_pcpu_id() == BSP_CPU_ID, "");
-
-		init_software_sram(true);
-	} else {
-		pr_dbg("Core %hu is up", pcpu_id);
-
-		pr_warn("Skipping VM configuration check which should be done before building HV binary.");
-
-		init_software_sram(false);
-
-		/* Initialize secondary processor interrupts. */
-		init_interrupt(pcpu_id);
-
-		timer_init();
-		ptdev_init();
-
-		/* Wait for boot processor to signal all secondary cores to continue */
-		wait_sync_change(&pcpu_sync, 0UL);
-	}
-
-	init_sched(pcpu_id);
-
-#ifdef CONFIG_RDT_ENABLED
-	setup_clos(pcpu_id);
-#endif
-
-	enable_smep();
-
-	enable_smap();
-}
-
-static uint16_t get_pcpu_id_from_lapic_id(uint32_t lapic_id)
-{
-	uint16_t i;
-	uint16_t pcpu_id = INVALID_CPU_ID;
-
-	for (i = 0U; i < phys_cpu_num; i++) {
-		if (per_cpu(lapic_id, i) == lapic_id) {
-			pcpu_id = i;
-			break;
-		}
-	}
-
-	return pcpu_id;
-}
-
 static void start_pcpu(uint16_t pcpu_id)
 {
 	uint32_t timeout;
+
+	if (startup_paddr == 0UL) {
+		startup_paddr = prepare_trampoline();
+	}
 
 	/* Update the stack for pcpu */
 	stac();
@@ -379,6 +156,11 @@ bool start_pcpus(uint64_t mask)
 	pcpu_sync = 0UL;
 
 	return ((pcpu_active_bitmap & mask) == mask);
+}
+
+void wait_all_pcpus_run(void)
+{
+	wait_sync_change(&pcpu_sync, 0UL);
 }
 
 void make_pcpu_offline(uint16_t pcpu_id)
@@ -469,14 +251,6 @@ static void set_current_pcpu_id(uint16_t pcpu_id)
 	msr_write(MSR_IA32_TSC_AUX, (uint64_t) pcpu_id);
 }
 
-static void print_hv_banner(void)
-{
-	const char *boot_msg = "ACRN Hypervisor\n\r";
-
-	/* Print the boot message */
-	printf(boot_msg);
-}
-
 static
 inline void asm_monitor(volatile const uint64_t *addr, uint64_t ecx, uint64_t edx)
 {
@@ -507,7 +281,7 @@ void wait_sync_change(volatile const uint64_t *sync, uint64_t wake_sync)
 	}
 }
 
-static void init_pcpu_xsave(void)
+void init_pcpu_xsave(void)
 {
 	uint64_t val64;
 	struct cpuinfo_x86 *cpu_info;
